# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Fingarden API Docs",
    "version": "1.0.0",
    "title": "Fingarden"
  },
  "basePath": "/",
  "paths": {
...
    "/products": {
      "get": {
        "tags": ["product"],
        "summary": "Get array of products",
        "description": "",
        "operationId": "getProducts",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Filtering products by name",
            "required": false,
            "type": "string"
          },
          {
            "name": "minPrice",
            "in": "query",
            "description": "Filtering products by price greater then minPrice",
            "required": false,
            "type": "integer"
          },
          {
            "name": "maxPrice",
            "in": "query",
            "description": "Filtering products by price less then maxPrice",
            "required": false,
            "type": "integer"
          },
          {
            "name": "desc",
            "in": "query",
            "description": "Filtering products by desc",
            "required": false,
            "type": "string"
          },
          {
            "name": "available",
            "in": "query",
            "description": "Filtering products by available",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "colors",
            "in": "query",
            "description": "Filtering products by colors URLs in array",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "categories",
            "in": "query",
            "description": "Filtering products by categories URLs in array",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "brands",
            "in": "query",
            "description": "Filtering products by brands names in array",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "description": "Filtering products by tags URLs in array",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "sizes",
            "in": "query",
            "description": "Filtering products by sizes URLs in array",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "parent",
            "in": "query",
            "description": "Filtering products by parent url",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "default": "name",
            "required": false,
            "type": "string"
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Ordering results. ASC | DESC",
            "default": "DESC",
            "required": false,
            "type": "string"
          },
          {
            "type": "string",
            "in": "query",
            "name": "offset",
            "required": false,
            "default": 0,
            "description": "qty rows to skip for pagination"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "qty rows to take for pagination",
            "default": 10,
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductResponse"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["product"],
        "summary": "Create a product",
        "description": "",
        "operationId": "createProduct",
        "produces": ["application/json"],
        "parameters": [
          {
            "type": "string",
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "admin-token"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/ProductDTO"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful created",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Product request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/sizes": {
      "get": {
        "tags": ["size"],
        "summary": "Get array of sizes",
        "description": "",
        "operationId": "getSizes",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "type": "string",
            "in": "query",
            "name": "name",
            "required": false
          },
          {
            "type": "string",
            "in": "query",
            "name": "url",
            "required": false
          },
          {
            "type": "string",
            "in": "query",
            "name": "parent",
            "required": false
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "in": "query",
            "name": "products",
            "required": false
          },
          {
            "type": "string",
            "in": "query",
            "name": "sortBy",
            "required": false,
            "default": "name"
          },
          {
            "type": "string",
            "in": "query",
            "name": "orderBy",
            "required": false,
            "default": "DESC",
            "description": "can be ASC or DESC"
          },
          {
            "type": "string",
            "in": "query",
            "name": "offset",
            "required": false,
            "default": 0,
            "description": "qty rows to skip for pagination"
          },
          {
            "type": "string",
            "in": "query",
            "name": "limit",
            "required": false,
            "default": 10,
            "description": "qty rows to take for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SizeResponse"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["size"],
        "summary": "Create a size",
        "description": "",
        "operationId": "createSize",
        "produces": ["application/json"],
        "parameters": [
          {
            "type": "string",
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "admin-token"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Size"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful created",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Size request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/sizes/{sizeId}": {
      "get": {
        "tags": ["size"],
        "summary": "Find size by id",
        "operationId": "findSizeById",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "sizeId",
            "in": "path",
            "description": "ID of size to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Size"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Size not found"
          }
        }
      },
      "put": {
        "tags": ["size"],
        "summary": "Update the size",
        "description": "",
        "operationId": "updateSize",
        "produces": ["application/json"],
        "parameters": [
          {
            "type": "string",
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "admin-token"
          },
          {
            "name": "sizeId",
            "in": "path",
            "description": "ID of size to return",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful updated",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/SizeWithoutJoins"
            }
          },
          "400": {
            "description": "Invalid Size request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Size not found"
          }
        }
      },
      "delete": {
        "tags": ["size"],
        "summary": "Delete the size",
        "description": "",
        "operationId": "deleteSize",
        "produces": ["application/json"],
        "parameters": [
          {
            "type": "string",
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "admin-token"
          },
          {
            "name": "sizeId",
            "in": "path",
            "description": "ID of tag to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful deleted",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "url": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Size not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/api/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "Brand": {
      "required": ["name", "url"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "showOnMain": {
          "type": "boolean",
          "default": false,
          "description": "value true can be post only with admin token"
        }
      },
      "xml": {
        "name": "Brand"
      }
    },
...
    "Category": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        },
        "parent": {
          "type": "object",
          "$ref": "#/definitions/Category"
        },
        "children": {
          "type": "object",
          "$ref": "#/definitions/Category"
        },
        "parameters": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/Parameter"
          }
        },
        "url": {
          "type": "string"
        }
      }
    },
...
    "Color": {
      "required": ["name", "url", "code"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "products": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/ProductWithoutJoins"
          },
          "readOnly": true
        },
        "url": {
          "type": "string"
        },
        "code": {
          "type": "string"
        }
      }
    },
...
    "Parameter": {
      "required": ["name"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        }
      }
    },
    "Rating": {
      "properties": {
        "1": {
          "type": "number"
        },
        "2": {
          "type": "number"
        },
        "3": {
          "type": "number"
        },
        "4": {
          "type": "number"
        },
        "5": {
          "type": "number"
        },
        "avg": {
          "type": "number"
        }
      }
    },
    "ParameterProduct": {
      "properties": {
        "productId": {
          "type": "string"
        },
        "parameterId": {
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "parameter": {
          "$ref": "#/definitions/Parameter"
        }
      }
    },
    "ParameterProductsDTO": {
      "required": ["parameterId", "value"],
      "properties": {
        "parameterId": {
          "type": "number"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "Product": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "shortDesc": {
          "type": "string"
        },
        "keywords": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        },
        "category": {
          "type": "object",
          "$ref": "#/definitions/Category"
        },
        "brand": {
          "type": "object",
          "$ref": "#/definitions/Brand"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/TagWithoutJoins"
          }
        },
        "sizes": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/SizeWithoutJoins"
          }
        },
        "url": {
          "type": "string"
        },
        "rating": {
          "$ref": "#/definitions/Rating"
        },
        "reviews": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReviewWithoutJoins"
          }
        },
        "questions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Question"
          }
        },
        "parameterProducts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ParameterProduct"
          }
        },
        "productVariants": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductVariant"
          }
        }
      }
    },
    "ProductResponse": {
      "properties": {
        "rows": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product"
          }
        },
        "length": {
          "type": "number"
        }
      }
    },
    "ProductWithoutJoins": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        },
        "url": {
          "type": "string"
        }
      }
    },
    "ProductVariant": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "price": {
          "type": "integer",
          "minimum": 1
        },
        "oldPrice": {
          "type": "integer"
        },
        "wholeSalePrice": {
          "type": "integer"
        },
        "available": {
          "type": "boolean"
        },
        "images": {
          "type": "string"
        },
        "color": {
          "type": "object",
          "$ref": "#/definitions/Color"
        }
      }
    },
    "ProductDTO": {
      "required": ["name", "category", "brand", "url"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "shortDesc": {
          "type": "string"
        },
        "keywords": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "array of tags IDs"
        },
        "category": {
          "type": "string",
          "description": "Category ID"
        },
        "brand": {
          "type": "string",
          "description": "Brand ID"
        },
        "url": {
          "type": "string"
        },
        "parameterProducts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ParameterProductsDTO"
          }
        },
        "productVariants": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductVariant"
          }
        }
      }
    },
...
    "TagWithoutJoins": {
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
    "Size": {
      "required": ["name", "url"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "products": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/ProductWithoutJoins"
          },
          "readOnly": true
        },
        "url": {
          "type": "string"
        }
      }
    },
    "SizeResponse": {
      "properties": {
        "rows": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Size"
          }
        },
        "length": {
          "type": "number"
        }
      }
    },
    "SizeWithoutJoins": {
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
...
    "Review": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "rating": {
          "type": "integer",
          "minimum": 1,
          "maximum": 10
        },
        "text": {
          "type": "string"
        },
        "images": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "showOnMain": {
          "type": "boolean",
          "default": false
        },
        "product": {
          "type": "object",
          "$ref": "#/definitions/Product"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comment"
          }
        },
        "reactions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReviewReaction"
          }
        }
      }
    },
    "Question": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "text": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "product": {
          "type": "object",
          "$ref": "#/definitions/Product"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/QuestionComment"
          }
        },
        "reactions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReactionQuestion"
          }
        }
      }
    },
...
    "ReviewWithoutJoins": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "rating": {
          "type": "integer",
          "minimum": 1,
          "maximum": 10
        },
        "text": {
          "type": "string"
        },
        "images": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "showOnMain": {
          "type": "boolean",
          "default": false
        },
        "productId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comment"
          }
        },
        "reactions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReviewReaction"
          }
        }
      }
    },
...
    "ReviewReaction": {
      "properties": {
        "id": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "reviewId": {
          "type": "string"
        },
        "reaction": {
          "type": "string",
          "default": "Like"
        }
      }
    },
    "ReactionQuestion": {
      "properties": {
        "id": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "questionId": {
          "type": "string"
        },
        "reaction": {
          "type": "string",
          "default": "Like"
        }
      }
    },
    "User": {
      "required": ["firstName", "lastName", "email", "password"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "isVerified": {
          "type": "boolean",
          "default": false
        },
        "role": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "Comment": {
      "properties": {
        "id": {
          "type": "string"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "review": {
          "type": "object",
          "$ref": "#/definitions/Review"
        },
        "text": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "reactions": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/CommentReaction"
          }
        }
      }
    },
    "QuestionComment": {
      "properties": {
        "id": {
          "type": "string"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "question": {
          "type": "object",
          "$ref": "#/definitions/Question"
        },
        "text": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "reactions": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/CommentReaction"
          }
        }
      }
    },
    "CommentReaction": {
      "properties": {
        "id": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "commentId": {
          "type": "string"
        },
        "reaction": {
          "type": "string",
          "default": "Like"
        }
      }
    },
...
  }
}